moved get and post functions into seperate files
