--- bestpass/public/index.html
@@ -1,155 +1,0 @@
-<!DOCTYPE html>
-<html lang="en" data-theme="dark">
-  <head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <meta http-equiv="X-UA-Compatible" content="ie=edge">
-
-    <title>BestPass</title>
-    <link rel="icon" type="image/x-icon" href="/images/favicon.ico">
-
-    <link
-        href="https://cdn.jsdelivr.net/npm/daisyui@5"
-        rel="stylesheet"
-        type="text/css"
-    />
-    <script src="https://cdn.jsdelivr.net/npm/@tailwindcss/browser@4"></script>
-    <script src="https://unpkg.com/htmx.org@2.0.4"></script>
-  </head>
-
-  <body class="bg-base-200">
-    <!-- Nav bar -->
-  <div class="navbar bg-base-100 shadow-sm">
-    <div class="flex-1">
-      <a class="btn btn-ghost text-xl">BestPass</a>
-    </div>
-    <div class="flex gap-2">
-      <div class="dropdown dropdown-end">
-        <div tabindex="0" role="button" class="btn btn-ghost btn-circle avatar">
-          <div class="w-10">
-            <img
-              alt="Tailwind CSS Navbar component"
-              src="./images/lock.png">
-          </div>
-        </div>
-        <ul
-          tabindex="0"
-          class="menu menu-sm dropdown-content bg-base-100 rounded-box z-1 mt-3 w-52 p-2 shadow">
-          <li hx-get="/api/username" hx-trigger="load" hx-swap="innerHTML"> Loading...></li>
-          <li><a>Settings</a></li>
-          <li><a href="/passwords">Login</a></li>
-          <li><a>Logout</a></li>
-        </ul>
-      </div>
-    </div>
-  </div>
-
-    <!-- Hero Section -->
-    <section class="hero min-h-screen bg-base-100">
-      <div class="hero-content text-center">
-        <div class="max-w-md">
-          <h1 class="text-5xl font-bold">Secure Your Passwords</h1>
-          <p class="py-6">A simple and secure way to manage all your passwords in one place.</p>
-          <button class="btn btn-primary">Get Started</button>
-        </div>
-      </div>
-    </section>
-  
-    <!-- Features Section -->
-    <section class="py-20 bg-base-100">
-      <div class="container mx-auto px-4">
-        <h2 class="text-3xl font-bold text-center mb-12">Features</h2>
-        <div class="grid grid-cols-1 md:grid-cols-3 gap-8">
-          <div class="card bg-base-200 shadow-xl">
-            <div class="card-body">
-              <h3 class="card-title">Secure Encryption</h3>
-              <p>Your passwords are encrypted with the latest security standards.</p>
-            </div>
-          </div>
-          <div class="card bg-base-200 shadow-xl">
-            <div class="card-body">
-              <h3 class="card-title">Cross-Platform Sync</h3>
-              <p>Access your passwords from any device, anywhere.</p>
-            </div>
-          </div>
-          <div class="card bg-base-200 shadow-xl">
-            <div class="card-body">
-              <h3 class="card-title">Easy to Use</h3>
-              <p>Simple and intuitive interface for managing your passwords.</p>
-            </div>
-          </div>
-        </div>
-      </div>
-    </section>
-  
-    <!-- Reviews Section -->
-    <section class="py-20 bg-base-100">
-      <div class="container mx-auto px-4">
-        <h2 class="text-3xl font-bold text-center mb-12">What Our Users Say</h2>
-        <div class="grid grid-cols-1 md:grid-cols-3 gap-8">
-          <!-- Review 1 -->
-          <div class="card bg-base-200 shadow-xl">
-            <div class="card-body">
-              <div class="flex items-center space-x-4">
-                <div class="avatar">
-                  <div class="w-12 h-12 rounded-full">
-                    <img src="./images/user.png" alt="User Avatar">
-                  </div>
-                </div>
-                <div>
-                  <h3 class="card-title">John Doe</h3>
-                  <p class="text-sm text-gray-500">Software Engineer</p>
-                </div>
-              </div>
-              <p class="mt-4">"This password manager has made my life so much easier. Highly recommended!"</p>
-            </div>
-          </div>
-  
-          <!-- Review 2 -->
-          <div class="card bg-base-200 shadow-xl">
-            <div class="card-body">
-              <div class="flex items-center space-x-4">
-                <div class="avatar">
-                  <div class="w-12 h-12 rounded-full">
-                    <img src="./images/user.png" alt="User Avatar">
-                  </div>
-                </div>
-                <div>
-                  <h3 class="card-title">Jane Smith</h3>
-                  <p class="text-sm text-gray-500">Product Manager</p>
-                </div>
-              </div>
-              <p class="mt-4">"I love how easy it is to sync my passwords across all my devices."</p>
-            </div>
-          </div>
-  
-          <!-- Review 3 -->
-          <div class="card bg-base-200 shadow-xl">
-            <div class="card-body">
-              <div class="flex items-center space-x-4">
-                <div class="avatar">
-                  <div class="w-12 h-12 rounded-full">
-                    <img src="./images/user.png" alt="User Avatar">
-                  </div>
-                </div>
-                <div>
-                  <h3 class="card-title">Mike Johnson</h3>
-                  <p class="text-sm text-gray-500">Freelancer</p>
-                </div>
-              </div>
-              <p class="mt-4">"The security features are top-notch. I feel safe using this password manager."</p>
-            </div>
-          </div>
-        </div>
-      </div>
-    </section>
-  
-    <!-- Footer -->
-    <footer class="footer footer-center p-10 bg-base-300 text-base-content">
-      <div>
-        <p>© 2023 Password Manager. All rights reserved.</p>
-      </div>
-    </footer>
-  
-  </body>
-</html>
