+++ bestpass/public/login.html
@@ -10,7 +10,7 @@
 <div class="relative flex flex-col justify-center h-screen overflow-hidden">
     <div class="w-full p-6 m-auto bg-white rounded-md shadow-md lg:max-w-lg">
         <h1 class="text-3xl font-semibold text-center text-purple-700">BestPass</h1>
-        <form class="space-y-4" action="/login" method="post">
+        <form class="space-y-4" action="/api/login" method="post">
             <div>
                 <label class="label">
                     <span class="text-base label-text">Email</span>
